# schema/openapi.yaml
openapi: 3.0.3
info:
  title: Agenty API
  version: v1
  description: |
    The API of an agenty server: it stores harnesses, runs them, and streams
    each run's events, approval requests included.

    Every request signs in with a user's bearer token. Harnesses and runs
    live in workspaces; a workspace the user is not a member of is not found,
    and neither is a run asked for in another workspace than its own. The
    exception: a user still reads the runs of conversations they started in
    a workspace they left, read-only.

    Errors are returned as an Error: 400 for an invalid request, 401 without
    a valid token, 403 for a request for a host that is not local, from a
    web page of an origin that is not configured, or for the audit log's
    runs, events and export by a user who is not an auditor, 404 for anything not
    found, workspaces included, 409 for cancelling a run that has finished,
    for following up a run that has not or that is followed up already, or
    for a run that has not finished but that this server does not run, 415
    for a body that is not JSON, 500 when the server fails, and 503 while
    the server is stopping.
    A run that cannot start, such as one whose grants no MCP server serves,
    is queued all the same and fails.

    This file is the API's source of truth: the Go types in internal/api and
    the server interface in internal/server are generated from it.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
...
  /v1/workspaces/{workspace}/runs/{id}/events:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: streamRunEvents
      summary: A run's events, as server-sent events
      description: |
        The events from the run's start, or after the one Last-Event-ID
        names, read from the audit log: its audit records, the approval
        requests it made, as they are now, and its end. The stream of a run
        that has not finished, queued, waiting or running, replays them,
        then follows the run until it finishes, or until the server stops,
        which leaves queued and waiting runs to the next. Each event has
        its id in the audit log as its SSE id, and its data is JSON:

        - `audit`: an AuditRecord
        - `approval`: an ApprovalRequest
        - `finished`: the Run, always the last event
      parameters:
        - name: Last-Event-ID
          in: header
          description: The id of the last event the reader saw; the stream goes on after it.
          schema: {type: integer, format: int64, minimum: 0}
      responses:
        "200":
          description: The event stream.
          content:
            text/event-stream:
              schema: {type: string}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Workspace:
      name: workspace
      in: path
      required: true
      schema: {type: string}
    RunID:
      name: id
      in: path
      required: true
      description: |
        A run of a conversation the user started. Any other run, also one of
        another member's conversation, is not found.
      schema: {type: string}
...
  responses:
    Error:
      description: An error.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
...